fix(front-end): added search function
Added search function and connected to anand's code
