+++ projects/front/components/trades-page/index.ts
@@ -1,6 +1,7 @@
 export * from './search';
 export * from './trading-plan';
+export * from './plan';
 export * from './types';
-export * from './temp-data';
+export * from './recent-trade';
 export * from './recent-trades';
 export * from './delete-modal';
