+++ projects/front/components/trades-page/search.tsx
@@ -1,7 +1,20 @@
 import Image from 'next/image';
 import { Box, Input } from '..';
+import { Dispatch, SetStateAction } from 'react';
 
-export const SearchInput = () => {
+type SearchType = {
+  value: string;
+  setValue: Dispatch<SetStateAction<string>>;
+  placeholder: string;
+  rightElement?: React.ReactNode;
+};
+
+export const SearchInput: React.FC<SearchType> = ({
+  value,
+  setValue,
+  placeholder,
+  rightElement,
+}) => {
   return (
     <Box className="bg-white h-14 px-10  py-3 flex items-center rounded-3xl w-full shadow-sm">
       <Box className="w-full flex items-center gap-4">
@@ -11,12 +24,14 @@ export const SearchInput = () => {
           height={20}
           width={20}
         />
-        <Input placeholder="Search" className="font-medium w-full" />
+        <Input
+          value={value}
+          onChange={e => setValue(e.target.value)}
+          placeholder={placeholder}
+          className="font-medium w-full"
+        />
       </Box>
-      <button className="bg-dark text-white flex-shrink-0 font-medium rounded-full p-2 px-6 flex gap-1">
-        <Image src="/icons/add.svg" height={22} width={22} alt="add icon" />
-        New Plan
-      </button>
+      {rightElement && rightElement}
     </Box>
   );
 };
